+++ module4/MOD.h
@@ -10,7 +10,7 @@ void mod_arr(unsigned int *arr1, unsigned int *arr2, unsigned int *Result,
 	const unsigned int thread_idx = (blockIdx.x * blockDim.x) + threadIdx.x;
 
     if(arr2[thread_idx] != 0)
-	    Result[thread_idx] = arr1[thread_idx] % arr2[thread_idx];
+	    Result[thread_idx] = static_cast<int>(arr1[thread_idx] % arr2[thread_idx]);
     else
         Result[thread_idx] = 99999;
 	
@@ -21,6 +21,7 @@ void mod_arr(unsigned int *arr1, unsigned int *arr2, unsigned int *Result,
 void Topmod(unsigned int *gpu_arr1, unsigned int *gpu_arr2,unsigned int num_blocks, 
               unsigned int num_threads, RESULT *finalResult)
 {
+    //Preparation to do the modulus in the kernel
 	const unsigned int ARRAY_SIZE     = num_blocks * num_threads;
 	unsigned int ARRAY_SIZE_IN_BYTES  = (sizeof(unsigned int) * (ARRAY_SIZE));
 	unsigned int cpu_Result[ARRAY_SIZE];
